Fix #181 and #188: Better pauli nodes recognition
This commit introduces a new method `is_pauli` on `M` commands, which
returns a pair `(axis, sign)` if the measure is Pauli, and `None`
otherwise.

This method takes an optional parameter `precision`
($\rho$, default: `1e-6`):
a measure is considered to be Pauli if the angle is in an interval
$[- \rho + k \cdot \frac \pi 2; \rho + k \cdot \frac \pi 2]$,
addressing the precision problem raised in #181.

`axis` is computed by using the `cos` and `sin` definitions of
`pauli.py`, solving the inconsistency bug reported in #188.

This commit introduces a new `Sign` class to make code manipulating
`Plus` and `Minus` to be more readable than when using `bool`.